windowsのビルドに成功

+++ src/takeyariGUI.js
@@ -2,9 +2,6 @@ const path = require("path");
 const fs = require("fs-extra");
 const exec = require("child_process");
 const util = require("./modules/util");
-const nil = require("./modules/nodeinstaller/linux");
-const niw = require("./modules/nodeinstaller/windows");
-
 var getProjectDirectory = function() {
   return path.join(process.env[(process.platform == "win32") ? "USERPROFILE" : "HOME"],"TakeyariViewer");
 };
@@ -17,7 +14,14 @@ var createProject = function(){
   if(projectName && (projectName.length > 0)) {
     let projectPath= path.join(getProjectDirectory(),projectName);
     fs.mkdirsSync(getProjectDirectory());
-    fs.copySync(path.join(util.getResourcePath(),'TakeyariViewer'),projectPath);
+    if(fs.existsSync(projectPath)) {
+        put("project already exist.");
+    } else {
+      fs.copySync(path.join(util.getResourcePath(),'TakeyariViewer'),projectPath);
+      if(fs.existsSync(projectPath)) {
+        put("project create success!");
+      }
+    }
   }
 };
 var buildProject = function() {
@@ -28,10 +32,12 @@ var buildProject = function() {
     if(process.platform === "linux") {
       buildLinux(projectPath);
     } else if(process.platform === "win32") {
+      buildWindows(projectPath);
     }
   }
 };
 var buildLinux = function(projectPath) {
+  const nil = require("./modules/nodeinstaller/linux");
   nil.nodeInstallProject(util.getResourcePath(), util.getResourcePath() ,projectPath, function(result) {
     if(result) {
       let buildingProc = exec.spawn(path.join(projectPath,"build.sh"));
@@ -44,17 +50,19 @@ var buildLinux = function(projectPath) {
   });
 };
 var buildWindows = function(projectPath) {
+  const niw = require("./modules/nodeinstaller/windows");
   niw.nodeInstallProject(util.getResourcePath(), util.getResourcePath() ,projectPath, function(result) {
     if(result) {
       put("downloading dependencies package.");
-      let npmInstallProc = exec.spawn(path.join("npm"),["install"],{cwd:projectPath});
+      let originPath = process.env["PATH"];
+      let npmInstallProc = exec.spawn("cmd",["/c",path.join("bin","node.exe"),path.join("bin","node_modules","npm","bin","npm-cli.js"),"install"],{cwd:projectPath});//パスを通し、文字コードをutf-8に変更する必要がある
       setMessageProc(npmInstallProc,function(){
         put("building portfolio site.");
-        let tscProc = exec.spawn(path.join("node_modules",".bin","tsc"),{cwd:projectPath});
+        let tscProc = exec.spawn("cmd",["/c",path.join("bin","node.exe"),path.join("node_modules","typescript","bin","tsc")],{cwd:projectPath});
         setMessageProc(tscProc,function() {
-          let nodeProc = exec.spawn(path.join("node"),[path.join("store_bundle","store_src","main.js")],{cwd:projectPath});
+          let nodeProc = exec.spawn("cmd",["/c",path.join("bin","node.exe"),path.join("store_bundle","store_src","main.js")],{cwd:projectPath});
           setMessageProc(nodeProc,function(){
-            let ngBuildProc = exec.spawn(path.join("node_modules",".bin","ng"),["build","--prod"],{cwd:projectPath});
+            let ngBuildProc = exec.spawn("cmd",["/c",path.join("bin","node.exe"),path.join("node_modules","@angular","cli","bin","ng"),"build","--prod"],{cwd:projectPath});
             setMessageProc(ngBuildProc,function(){
               put("complete building.");
               put("build end!");
@@ -65,6 +73,8 @@ var buildWindows = function(projectPath) {
     } else {
       put("failed to provide tool for build.");
     }
+  },function(message){
+    put(message);
   });
 };
 var setMessageProc = function(proc,callback){
